Add copy() method to Profile and its descendants (Prerequisite for #78)
The Profile.copy() method allows creation of a deep copy. This decouples
the profile publication from processing on the output (ui, console, ...)
side.

This means for instance that filters in the UI can be applied on a copy,
and when the filter is removed, the original profile can be reinstated.

The copy also can fulfill the role of a snapshot (i.e. for implementing
#78).