Fix Monitor actions and fix filter equals
The VirtualMachine userDir does not necessarily point to the directory
which was specified with the logPath option for the agentpath vm
argument. Therefore, now the vmargs are parsed, the logpath argument
value is extracted and a FileLogSource based on a File with that
argument as path can be gotten from the VirtualMachine.

I noticed belatedly I broke the fix to the numeric filter equals()
method (which returned true on double inequality rather than equality).
So, that fix has been applied again.