VERY IMPORTANT PERFORMANCE UPDATE : Decoupled Views from ProfileListener
There was a serious performance issue with large logs. The UI would hang
or even OOM.

One of the underlying causes was that every view (Flat, Tree,
FlameGraph) was a Listener, and they all would be updated when a new
Profile was selected.

In case of the FlameGraph View, the performance hit is extremely large,
and the UI would hang.

Now, the views have been decoupled. The ProfileContext containing an
ObjectProperty<Profile> is created first, and the different views will
bind or unbind their own ObjectProperty<Profile> to that when they are
shown or hidden.

As a result, the rendition for each view only takes place at the time
they are shown, and even then only if the Profile changes. When they are
hidden, they no longer react to new Profiles emitted by the back end
until such a time as they are shown again.