Added Freeze function, fix for #78
The Freeze function, activated with a button (clock icon), makes the UI
stop updating the live profile. The incoming changes are still being
stored, the UI just doesn't react to it any more.

When the profile is unfrozen, the UI starts updating again and jumps to
the most recently received profile information.