Moved column configuration higher up & renamed quite a few fields for consistency.
There seems to be an issue with the comparison TreeTable context menu
which only shows up sporadically. Tried to fix it but so far no luck.
Will investigate later.

This version counts as the "milestone" I've talked about in the PR
comments. No major or even minor changes will be added to this PR
anymore. At most very small fixes if/when requested by reviewers.