Re-added gradle build scripts etc in new branch
The idea of this branch is first off to have the gradle build resources
available for those who need it (e.g. me).

Also, there are some features gradle offers which I'm not so sure about
in the maven case. Because of these I'd like to make the case in a
future PR for moving from maven to gradle.