Work In Progress : New Lean Profile + Aggregation framework
Commit + push to safeguard work...

This is the groundwork for more flexible aggregation. The LeanProfile
contains only the basic profile data without any redundancy. It only
counts and add times for the frames in the stack tree.

The aggregation package provides extra linking for easy aggregation over
threads, and for backtrace access (i.e. given a method name, find all
stacks which call this method).

Also there are generalized AggregatedEntries which will be the input for
the controllers.

The first couple of aggregators are still being written.