Refactored names & cache aggregations in AggrProfile
The objects in the aggregation package had names which made for hard
reading. As a left-to-right reader, if the first portion of the name is
always the same (and it's +- in the package name anyway) and the last
portion not distinct enough, it becomes straining to read the code using
these types.

Also, less broken-up lines is a Good Thing(tm) IMHO.