Got rid of unneeded extra thread (moderator) for live profiling
The front-end requests profiles from the Collector, which delivers them
to a listener. That listener transforms it into an AggregationProfile in
a background task, which effectively decouples the threads.

The pull-mechanism ensures we don't continuously have to emit profiles,
they are only emitted at the rate set by the UI (which also means we can
change the emission rate from the front-end, once the control is
implemented). This saves a lot of processing by avoiding unnecessary
profile object creation on emission (which entails a copy of the trees).