Move descendants view out of TreeView, up one level
Needed for reusability of TreeView, which will be used to show "called
by" and "calling" methods for items selected in FlatView. But there we
don't need the descendants view included.

Put otherwise, TreeView and Flatview now pretty much are reusable
components with the single specific purpose of showing a Tree and Flat
aggregation respectively.