Added Called By & Calling Tree Views To Flat View
When clicking on an entry in the flat aggregation, you now get at the
bottom the calling and called methods, aggregated by FQMN.

Still needs verification and cleanup, but the base flow and aggregation
is there.

Also, some trivial cleanup, and Thread LeanNodes are no longer in the
parent Set of the FqmnLinks.