Move reference down to entry-level
The "reference" concept implementation was a bit broken. The idea is
that an Entry calculates its percentage values respective to some
reference, which can be either the global aggregation of all entries, or
the thread-level aggregation, or the immediate parent (in case of
Nodes).

The original implementation set the reference at Aggregation-level,
which obviously meant that all Entries in the aggregation had the same
reference, so only the first mode could be supported.

Also included is a util class for switching references. Not used yet,
but it's there for exposing this in the front-end fairly soonish, if all
goes well.