Allow grouping selection in UI
This means that for the Flat and Tree views, you can now choose the way
nodes are aggregated : by thread id, name or over threads, and by Frame
method name, methodname + line nr or file + bci (to be reviewed, maybe
it has to be method + bci)

Currently the Ancestor/Descendants tree aggreagations at the bottom of
the Flat View are broken, but they were not quite correct anyway. Should
get fixed next week.