Fixed duplicate ancestor counting
Added a unique ID to LeanNode so we can weed out duplicates (like in the
aggregator fixed here), but it may also become useful in other cases
(e.g. in the future we can push out incremental LeanProfile updates
instead of full LeanProfiles which may (or may not) save a lot of
processing).

Also some very minor refactoring.