Fixed diff refresh
When the base or new changed in the diff, the old aggregations were not
cleared out, resulting in stale entries (e.g. when changing the frame
grouping).

Changed so that diff now requires base + new to be set simultaneously,
which makes everything simpler (originally though that they'd need to be
set separately sometimes but that was a mistaken assumption).