Added Package doc for core.aggregation + added By Method Id grouping
The documentation describes how all the aggregation data structures hang
together, how aggregations can be created and filtered etc.

And the Method Id grouping ? Well I accidentally added a line saying it
was supported in the documentation. While reviewing I figured out it
wasn't. But it was sooooo simple to add it... Couldn't resist it. Sorry.