Dialog code simplified a lot
While I don't plan on adding any features to the #184 PR, this is a
simplification which should actually make it easier for reviewers to
understand the code, since Dialogs and Dialog Controllers now are
handled much the same way as the other ViewControllers, using FXML
injection, instead of the roundabout FXMLLoader stuff and explicit
Dialog object configuration.

Up to now, I never found any resource on the web documenting how to
nicely do Dialogs this way. But I finally figured it out, after seeing a
couple of sentences buried in
https://docs.oracle.com/javase/8/javafx/api/javafx/fxml/doc-files/introduction_to_fxml.html
, specifically the remark about needing to nest it in an fx:define.

It doesn't help that SceneBuilder also offers no support for this
mechanism.

Anyway, instead of the ugly explicit FXMLLoader way which you can find
in almost any discussion of Dialog creation on the 'Net, which among
other things requires explicit FXML paths in the source code, here is
now what I strongly suspect is the 'proper' way to do it through FXML
controller injection. Yay. And I missed my train because of this.