Deep rework to simplify testing greatly
While writing the FX tests, it became apparent quickly that checking of
the test scenarios should be automated, for the tests to remain
maintainable.

The SimplifiedLogScenario now encapsulates the scenario, calculates the
aggregation result and provides automated checking.

Some tests now have been parameterized using the scenarios, and are as a
result immensely shorter.

I'm aware of course of the fact that the extra complexity in the test
itself, specifically in the SimplifiedLogScenario, now raises the
question of "who tests the tests" but the implementation of the
aggregation in there is sufficiently different from the aggregation
logic in the core.aggregation package to be fairly certain that the test
logic is valid.