Statically link libstdc++ to avoid ABI compatibility issues
This change allows the downloaded binaries to run on CentOS.

The agent depends on a C++ runtime, which, for GNU toolchain,
is provided by libstdc++. The default is to dynamically link
with that library. If the toolchain used for building
the agent is newer than the system-supplied libstdc++, a pre-built
agent cannot load. Such is the case when building on Ubuntu with GCC 5,
and trying to load the agent on CentOS 7, which is based on GCC 4.8.

This change statically links the C++ runtime, removing the chance for
C++ runtime mismatch. It increases the agent's size by 1.7MB, which is
about 100%. However, the increase is fixed (not supposed to further
increase should the agent become larger) and the agent is still reasonably
small in today's standards.